Move `grpcweb/common.go` into its own package
