+++ internal/api/common/common_test.go
@@ -1,4 +1,4 @@
-package grpcweb
+package common
 
 import (
 	"context"
