+++ internal/api/grpcweb/actioncacheproxy/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/buildbarn/bb-portal/internal/api/grpcweb/actioncacheproxy",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//internal/api/grpcweb",
+        "//internal/api/common",
         "@bazel_remote_apis//build/bazel/remote/execution/v2:remote_execution_go_proto",
         "@com_github_buildbarn_bb_storage//pkg/auth",
         "@org_golang_google_grpc//codes",
