+++ internal/api/grpcweb/buildqueuestateproxy/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/buildbarn/bb-portal/internal/api/grpcweb/buildqueuestateproxy",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//internal/api/grpcweb",
+        "//internal/api/common",
         "@com_github_buildbarn_bb_remote_execution//pkg/proto/buildqueuestate",
         "@com_github_buildbarn_bb_storage//pkg/auth",
         "@org_golang_google_grpc//codes",
